# optimize performance of the repeat function (up to 50% faster)