@zjregee 2025-02-16
Hi, @alamb.

In addition, thank you very much for your help, but I have tried some suggestions in #14610 to reduce memory copy by using the `write!`, but I have not achieved any results. I think there may be the following problems:
- By looking at the source code, the implementation of the `write!` seems to be the same as `append_value`, and is also implemented by calling `append_slice`. There is still a memory copy here, and using `write!` does not seem to have any special meaning.
- When using the following method for loop insertion, there seems to be no performance advantage when the number of loops is too large, because the `str::repeat` reduces the number of loops through exponential copying (although there are still more memory allocations here).
```Rust
for _ in 0..number {
    write!(builder, "{}", string)?;
}
builder.append_value("")
```