@alamb 2025-02-16
BTW I got a flamegraph of this run while using

```shell
andrewlamb@Andrews-MacBook-Pro-2:~/Software/datafusion$ sudo flamegraph -- target/release/deps/repeat-f851d6dc5c039e6c --bench times/repeat_large_string
```

Here it is in case you are interested:


<img width="1192" alt="Screenshot 2025-02-16 at 6 44 37 AM" src="https://github.com/user-attachments/assets/a2fc2e63-eefb-4026-8dd8-c2ac7b2054b9" />

While there is some malloc/free that shows up a lot of the time is spent copying the strings which makes sense